4BIT-1/SUI/Umela inteligence pro Valku kostek/xmorav41/supp-xmorav41/requirements.txt
hexutil
numpy
pyqt5==5.14
matplotlib
scipy


torchvision
sklearn
